Introduce version migration mechanism
