+++ migration/v1/model/application.go
@@ -1,0 +1,84 @@
+package model
+
+import (
+	"github.com/konveyor/tackle2-hub/model"
+)
+
+type Application struct {
+	Model
+	BucketOwner
+	Name              string `gorm:"index;unique;not null"`
+	Description       string
+	Review            *Review `gorm:"constraint:OnDelete:CASCADE"`
+	Repository        model.JSON
+	Binary            string
+	Facts             model.JSON
+	Comments          string
+	Tasks             []Task     `gorm:"constraint:OnDelete:CASCADE"`
+	Tags              []Tag      `gorm:"many2many:ApplicationTags;constraint:OnDelete:CASCADE"`
+	Identities        []Identity `gorm:"many2many:ApplicationIdentity;constraint:OnDelete:CASCADE"`
+	BusinessServiceID uint       `gorm:"index"`
+	BusinessService   *BusinessService
+}
+
+type Dependency struct {
+	Model
+	ToID   uint         `gorm:"index"`
+	To     *Application `gorm:"foreignKey:ToID;constraint:OnDelete:CASCADE"`
+	FromID uint         `gorm:"index"`
+	From   *Application `gorm:"foreignKey:FromID;constraint:OnDelete:CASCADE"`
+}
+
+type Review struct {
+	Model
+	BusinessCriticality uint   `gorm:"not null"`
+	EffortEstimate      string `gorm:"not null"`
+	ProposedAction      string `gorm:"not null"`
+	WorkPriority        uint   `gorm:"not null"`
+	Comments            string
+	ApplicationID       uint `gorm:"uniqueIndex"`
+	Application         *Application
+}
+
+type Import struct {
+	Model
+	Filename            string
+	ApplicationName     string
+	BusinessService     string
+	Comments            string
+	Dependency          string
+	DependencyDirection string
+	Description         string
+	ErrorMessage        string
+	IsValid             bool
+	RecordType1         string
+	ImportSummary       ImportSummary
+	ImportSummaryID     uint `gorm:"index"`
+	Processed           bool
+	ImportTags          []ImportTag `gorm:"constraint:OnDelete:CASCADE"`
+	BinaryGroup         string
+	BinaryArtifact      string
+	BinaryVersion       string
+	BinaryPackaging     string
+	RepositoryKind      string
+	RepositoryURL       string
+	RepositoryBranch    string
+	RepositoryPath      string
+}
+
+type ImportSummary struct {
+	Model
+	Content        []byte
+	Filename       string
+	ImportStatus   string
+	Imports        []Import `gorm:"constraint:OnDelete:CASCADE"`
+	CreateEntities bool
+}
+
+type ImportTag struct {
+	Model
+	Name     string
+	TagType  string
+	ImportID uint `gorm:"index"`
+	Import   *Import
+}
