+++ migration/v1/model/bucket.go
@@ -1,0 +1,5 @@
+package model
+
+type BucketOwner struct {
+	Bucket string `gorm:"index"`
+}
