+++ migration/v1/model/volume.go
@@ -1,0 +1,8 @@
+package model
+
+type Volume struct {
+	Model
+	Name     string `gorm:"<-:create;unique"`
+	Capacity string
+	Used     string
+}
